feat: add an entire new dice class based on Regex
This regex includes support for exploding dice and savage. It will be very simple to replace the format of the command: i.e. from current "1d4es" (1d4 exploding and savage) to "!1d4" (do not add savage dice)

This is a savage bot, so it should add the savage dice by default, also exploding. So the only necessary control should be for the addition of the savage dice